Project/apps/resource_server/pyproject.toml
[tool.poetry]
name = "resource-server"
version = "0.1.0"
description = ""
authors = [
    "André Costa Lima <up202008169@edu.fe.up.pt>",
]

[tool.poetry.dependencies]
python = "^3.12"
flask = { version = "^3.0.3", extras = ["async"] } 
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
python-ulid = "^3.0.0"
flask-sqlalchemy = "^3.1.1"
gunicorn = "^23.0.0"
secure-endpoints = {path = "../../packages/secure_endpoints", develop = true}
flask-pydantic = "^0.12.0"
utils = {path = "../../packages/utils", develop = true}

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[tool.poetry.scripts]
format = "black:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
